requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
django-filter==25.1
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
